requirements.txt
faster-whisper==1.1.1
googletrans==4.0.0rc1
pyttsx3==2.98
sounddevice==0.5.2
scipy==1.15.2
numpy==2.2.3
PyYAML==6.0.2
psutil==7.0.0
